+++ MOVEit.MIA/Private/Globals.ps1
@@ -13,6 +13,7 @@ $script:Context = @{
     Default = @{
         BaseUri = ''
         Token = @()
+        SkipCertificateCheck = $false
     }
 }